+++ libEM/emdata_transform.cpp
@@ -1894,7 +1894,7 @@ void EMData::insert_scaled_sum(EMData *block, const FloatPoint &center,
 		for (int x=x0; x<=x1; x++) {
 			for (int y=y0; y<=y1; y++) {
 				for (int z=z0; z<=z1; z++) {
-					idx = x + y * nx + z * nx * ny;
+					idx = x + y * nx + (size_t)z * nx * ny;
 					data[idx] +=
 						block->sget_value_at_interp((x-center[0])/scale+bx,(y-center[1])/scale+by,(z-center[2])/scale+bz);
 				}
